#1 - using react-router-dom template
